chore: rename db_service package to dbservice
This is because staticcheck rule ST1003 recommends against having an
underscore in a package name (except for _test packages)

[#177206960](https://www.pivotaltracker.com/story/show/177206960)
