+++ brokerapi/broker/last_instance_operation.go
@@ -4,10 +4,9 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/cloudfoundry/cloud-service-broker/db_service/models"
-	"github.com/cloudfoundry/cloud-service-broker/pkg/broker"
-
 	"code.cloudfoundry.org/lager"
+	"github.com/cloudfoundry/cloud-service-broker/dbservice/models"
+	"github.com/cloudfoundry/cloud-service-broker/pkg/broker"
 	"github.com/cloudfoundry/cloud-service-broker/utils/correlation"
 	"github.com/pivotal-cf/brokerapi/v8/domain"
 	"github.com/pivotal-cf/brokerapi/v8/domain/apiresponses"
