+++ brokerapi/broker/provision_test.go
@@ -6,7 +6,7 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/cloudfoundry/cloud-service-broker/db_service/models"
+	"github.com/cloudfoundry/cloud-service-broker/dbservice/models"
 
 	"github.com/cloudfoundry/cloud-service-broker/pkg/varcontext"
