+++ dbservice/setup_db.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package db_service
+package dbservice
 
 import (
 	"crypto/tls"
