+++ internal/encryption/delete_password_metadata.go
@@ -1,7 +1,7 @@
 package encryption
 
 import (
-	"github.com/cloudfoundry/cloud-service-broker/db_service/models"
+	"github.com/cloudfoundry/cloud-service-broker/dbservice/models"
 	"gorm.io/gorm"
 )
