+++ internal/encryption/passwordcombiner/combiner.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/cloudfoundry/cloud-service-broker/db_service/models"
+	"github.com/cloudfoundry/cloud-service-broker/dbservice/models"
 	"github.com/cloudfoundry/cloud-service-broker/internal/encryption/gcmencryptor"
 	"github.com/cloudfoundry/cloud-service-broker/internal/encryption/passwordparser"
 	"golang.org/x/crypto/pbkdf2"
