+++ internal/encryption/update_password_metadata.go
@@ -3,7 +3,7 @@ package encryption
 import (
 	"fmt"
 
-	"github.com/cloudfoundry/cloud-service-broker/db_service/models"
+	"github.com/cloudfoundry/cloud-service-broker/dbservice/models"
 	"gorm.io/gorm"
 )
