+++ internal/storage/update_all_records_test.go
@@ -1,7 +1,7 @@
 package storage_test
 
 import (
-	"github.com/cloudfoundry/cloud-service-broker/db_service/models"
+	"github.com/cloudfoundry/cloud-service-broker/dbservice/models"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
