+++ pkg/broker/service_provider.go
@@ -17,7 +17,7 @@ package broker
 import (
 	"context"
 
-	"github.com/cloudfoundry/cloud-service-broker/db_service/models"
+	"github.com/cloudfoundry/cloud-service-broker/dbservice/models"
 	"github.com/cloudfoundry/cloud-service-broker/internal/storage"
 	"github.com/cloudfoundry/cloud-service-broker/pkg/varcontext"
 	"github.com/pivotal-cf/brokerapi/v8/domain"
