+++ pkg/providers/tf/provider.go
@@ -23,7 +23,7 @@ import (
 	"github.com/cloudfoundry/cloud-service-broker/pkg/providers/tf/hclparser"
 
 	"code.cloudfoundry.org/lager"
-	"github.com/cloudfoundry/cloud-service-broker/db_service/models"
+	"github.com/cloudfoundry/cloud-service-broker/dbservice/models"
 	"github.com/cloudfoundry/cloud-service-broker/internal/storage"
 	"github.com/cloudfoundry/cloud-service-broker/pkg/broker"
 	"github.com/cloudfoundry/cloud-service-broker/pkg/providers/builtin/base"
